(refactor) add abstract services
