+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/service/AbstractChatService.java
@@ -1,0 +1,32 @@
+package ru.tinkoff.edu.java.scrapper.service;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.transaction.annotation.Transactional;
+import ru.tinkoff.edu.java.scrapper.model.Chat;
+import ru.tinkoff.edu.java.scrapper.repository.ChatRepository;
+
+
+@Slf4j
+public abstract class AbstractChatService implements ChatService {
+
+    private final ChatRepository chatRepository;
+
+    public AbstractChatService(ChatRepository chatRepository) {
+        this.chatRepository = chatRepository;
+    }
+
+    @Transactional
+    @Override
+    public void register(long chatId) {
+        if (chatRepository.findById(chatId).isEmpty()) {
+            chatRepository.save(new Chat().setId(chatId));
+            log.info("Chat with %s ID was registered".formatted(chatId));
+        }
+    }
+
+    @Transactional
+    @Override
+    public void unregister(long chatId) {
+        chatRepository.removeById(chatId);
+    }
+}
