+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/service/jooq/JooqChatService.java
@@ -1,31 +1,12 @@
 package ru.tinkoff.edu.java.scrapper.service.jooq;
 
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.transaction.annotation.Transactional;
-import ru.tinkoff.edu.java.scrapper.model.Chat;
 import ru.tinkoff.edu.java.scrapper.repository.jooq.JooqChatRepository;
-import ru.tinkoff.edu.java.scrapper.service.ChatService;
+import ru.tinkoff.edu.java.scrapper.service.AbstractChatService;
 
 
-@Slf4j
-@RequiredArgsConstructor
-public class JooqChatService implements ChatService {
+public class JooqChatService extends AbstractChatService {
 
-    private final JooqChatRepository chatRepository;
-
-    @Transactional
-    @Override
-    public void register(long chatId) {
-        if (chatRepository.findById(chatId).isEmpty()) {
-            chatRepository.save(new Chat().setId(chatId));
-            log.info("Chat with %s ID was registered".formatted(chatId));
-        }
-    }
-
-    @Transactional
-    @Override
-    public void unregister(long chatId) {
-        chatRepository.removeById(chatId);
+    public JooqChatService(JooqChatRepository chatRepository) {
+        super(chatRepository);
     }
 }
