+++ scrapper/src/test/java/util/JdbcIntegrationEnvironment.java
@@ -7,7 +7,7 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
 import ru.tinkoff.edu.java.scrapper.configuration.ApplicationProperties;
-import ru.tinkoff.edu.java.scrapper.configuration.database.JdbcAccessConfiguration;
+import ru.tinkoff.edu.java.scrapper.configuration.JdbcAccessConfiguration;
 
 
 @ContextConfiguration(classes = JdbcAccessConfiguration.class)
